requirements.txt
google-generativeai
gspread
oauth2client
streamlit
python-dotenv
pillow
